app/engine.py
"""Unified inference engine — loads all models once and provides
methods for ASR, LLM generation (blocking and streaming), and TTS synthesis.

Key optimization: generate_stream() yields audio chunks as the LLM
produces sentences, allowing the ESP32 to start playback while the
LLM is still generating. This cuts perceived latency by 40-60%.
"""

import io
import wave
import struct
import tempfile
import logging
import re
from typing import Iterator

from llama_cpp import Llama
from faster_whisper import WhisperModel

from .config import settings

logger = logging.getLogger(__name__)

# Regex to strip <think>...</think> blocks from Qwen 3 output
THINK_PATTERN = re.compile(r"<think>.*?</think>", re.DOTALL)

# Sentence boundary detection for Spanish — split on . ! ? ; and newlines
SENTENCE_END = re.compile(r'[.!?;。]\s*|\n')


class InferenceEngine:
    """Manages the ML models and exposes high-level inference methods.

    All models are loaded once at startup and reused across requests.
    Thread safety is handled at the FastAPI level with asyncio.Lock.
    """

    def __init__(self):
        self.llm: Llama | None = None
        self.whisper: WhisperModel | None = None
        self.tts_voice = None  # PiperVoice | None — imported conditionally
        self._ready = False

    def load_all(self, requires_tts: bool = True) -> None:
        """Load models into memory. Called once at server startup.

        Args:
            requires_tts: If False, skip loading Piper TTS to save ~60 MB
                          of RAM. Set automatically based on which pipelines
                          are active.
        """

        logger.info("Loading LLM: %s (threads=%d, ctx=%d, batch=%d)",
                     settings.MODEL_PATH, settings.N_THREADS,
                     settings.N_CTX, settings.N_BATCH)
        self.llm = Llama(
            model_path=settings.MODEL_PATH,
            n_threads=settings.N_THREADS,
            n_ctx=settings.N_CTX,
            n_batch=settings.N_BATCH,
            n_gpu_layers=0,
            verbose=False,
        )

        logger.info("Loading Whisper model: %s (lang=%s)",
                     settings.WHISPER_MODEL, settings.WHISPER_LANGUAGE)
        self.whisper = WhisperModel(
            settings.WHISPER_MODEL,
            device="cpu",
            compute_type="int8",
            cpu_threads=4,
        )

        if requires_tts:
            from piper.voice import PiperVoice

            logger.info("Loading Piper voice: %s", settings.PIPER_VOICE)
            self.tts_voice = PiperVoice.load(settings.PIPER_VOICE)
        else:
            logger.info("TTS not required by any pipeline — skipping Piper")

        self._ready = True
        logger.info("All required models loaded successfully")

    def is_ready(self) -> bool:
        return self._ready

    @property
    def tts_loaded(self) -> bool:
        return self.tts_voice is not None

    # ── ASR ─────────────────────────────────────────────────

    def transcribe(self, audio_bytes: bytes) -> str:
        """ASR: Convert audio bytes (WAV) to Spanish text."""
        with tempfile.NamedTemporaryFile(suffix=".wav", delete=True) as tmp:
            tmp.write(audio_bytes)
            tmp.flush()

            segments, info = self.whisper.transcribe(
                tmp.name,
                language=settings.WHISPER_LANGUAGE,
                beam_size=1,
                best_of=1,
                temperature=0.0,
                vad_filter=True,
                vad_parameters=dict(
                    min_silence_duration_ms=500,
                    speech_pad_ms=200,
                ),
                condition_on_previous_text=False,
                without_timestamps=True,
                word_timestamps=False,
            )
            return " ".join(seg.text.strip() for seg in segments)

    # ── LLM (blocking — for robot JSON mode) ───────────────

    def generate(self, messages: list[dict], json_mode: bool = False) -> str:
        """LLM: Generate a complete text response (non-streaming).
        Used for robot commands where we need the full JSON response
        before parsing.

        Args:
            messages: Chat messages including system prompt and history.
            json_mode: If True, constrain output to valid JSON.

        Returns:
            The generated text, with any <think> blocks stripped.
        """
        kwargs = dict(
            messages=messages,
            max_tokens=settings.MAX_TOKENS,
            temperature=settings.TEMPERATURE,
            top_p=0.8,
            top_k=20,
            presence_penalty=1.5,
        )

        if json_mode:
            kwargs["response_format"] = {"type": "json_object"}

        result = self.llm.create_chat_completion(**kwargs)
        content = result["choices"][0]["message"]["content"] or ""
        content = THINK_PATTERN.sub("", content).strip()
        return content

    # ── LLM (streaming — for assistant) ────────────────────

    def generate_stream(self, messages: list[dict]) -> Iterator[str]:
        """LLM: Stream tokens and yield complete sentences.

        Uses llama-cpp-python's stream=True which returns an Iterator
        of ChatCompletionChunk dicts. We accumulate tokens until we
        detect a sentence boundary, then yield the complete sentence.

        This allows the caller to synthesize audio for each sentence
        while the LLM continues generating the next one.

        Yields:
            Complete sentences as they are formed from streamed tokens.
        """
        stream = self.llm.create_chat_completion(
            messages=messages,
            max_tokens=settings.MAX_TOKENS,
            temperature=settings.TEMPERATURE,
            top_p=0.8,
            top_k=20,
            presence_penalty=1.5,
            stream=True,
        )

        buffer = ""
        in_think_block = False

        for chunk in stream:
            delta = chunk["choices"][0]["delta"]
            token = delta.get("content", "")
            if not token:
                continue

            # Track and skip <think>...</think> blocks
            buffer += token
            if "<think>" in buffer and not in_think_block:
                in_think_block = True
            if in_think_block:
                if "</think>" in buffer:
                    # Remove the entire think block
                    buffer = THINK_PATTERN.sub("", buffer)
                    in_think_block = False
                else:
                    continue  # Still inside think block, don't yield

            # Check for sentence boundary
            match = SENTENCE_END.search(buffer)
            if match:
                # Yield everything up to and including the sentence end
                end_pos = match.end()
                sentence = buffer[:end_pos].strip()
                buffer = buffer[end_pos:]
                if sentence:
                    yield sentence

        # Yield any remaining text
        remaining = buffer.strip()
        if remaining and not in_think_block:
            yield remaining

    # ── TTS ─────────────────────────────────────────────────

    def synthesize(self, text: str) -> bytes:
        """TTS: Convert text to WAV audio bytes (complete file)."""
        buffer = io.BytesIO()
        with wave.open(buffer, "wb") as wav_file:
            wav_file.setnchannels(1)
            wav_file.setsampwidth(2)
            wav_file.setframerate(self.tts_voice.config.sample_rate)
            self.tts_voice.synthesize(text, wav_file, sentence_silence=0.2)
        return buffer.getvalue()

    def synthesize_raw(self, text: str) -> bytes:
        """TTS: Convert text to raw PCM int16 bytes (no WAV header).
        Used for streaming — individual chunks that will be assembled
        into a WAV by the endpoint."""
        pcm_data = b""
        for audio_bytes in self.tts_voice.synthesize_stream_raw(text):
            pcm_data += audio_bytes
        return pcm_data

    def pcm_to_wav(self, pcm_chunks: list[bytes]) -> bytes:
        """Assemble raw PCM chunks into a complete WAV file."""
        all_pcm = b"".join(pcm_chunks)
        buffer = io.BytesIO()
        with wave.open(buffer, "wb") as wav_file:
            wav_file.setnchannels(1)
            wav_file.setsampwidth(2)
            wav_file.setframerate(self.tts_voice.config.sample_rate)
            wav_file.writeframes(all_pcm)
        return buffer.getvalue()
