requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
llama-cpp-python>=0.3.0
faster-whisper>=1.2.0
piper-tts>=1.4.0
huggingface-hub>=0.20.0
numpy>=1.24.0
